refactor: separates context elements in different files
