+++ src/contexts/TaskContext/TaskContextProvider.tsx
@@ -1,0 +1,19 @@
+import { useState } from 'react';
+import { initialTaskState } from './initialTaskState';
+import { TaskContext } from '.';
+
+// Context Provider & Props
+
+type TaskContentProviderProps = {
+  children: React.ReactNode;
+};
+
+export function TaskContextProvider({ children }: TaskContentProviderProps) {
+  const [state, setState] = useState(initialTaskState);
+
+  return (
+    <TaskContext.Provider value={{ state, setState }}>
+      {children}
+    </TaskContext.Provider>
+  );
+}
