+++ src/contexts/TaskContext/index.tsx
@@ -1,20 +1,6 @@
-import { createContext, useContext, useState } from 'react';
+import { createContext } from 'react';
 import type { TaskStateModel } from '../../models/TaskStateModel';
-
-// Initial value for TaskState
-
-const initialState: TaskStateModel = {
-  tasks: [],
-  secondsRemaining: 0,
-  formattedSecondsRemaining: '00:00',
-  activeTask: null,
-  currentCycle: 0,
-  config: {
-    workTime: 25,
-    shortBreakTime: 5,
-    longBreakTime: 15,
-  },
-};
+import { initialTaskState } from './initialTaskState';
 
 // Type Props & initial Props for Context
 
@@ -24,32 +10,10 @@ type TaskContextProps = {
 };
 
 const initialContextValue = {
-  state: initialState,
+  state: initialTaskState,
   setState: () => {},
 };
 
 // Context creation
 
 export const TaskContext = createContext<TaskContextProps>(initialContextValue);
-
-// Context Provider & Props
-
-type TaskContentProviderProps = {
-  children: React.ReactNode;
-};
-
-export function TaskContextProvider({ children }: TaskContentProviderProps) {
-  const [state, setState] = useState(initialState);
-
-  return (
-    <TaskContext.Provider value={{ state, setState }}>
-      {children}
-    </TaskContext.Provider>
-  );
-}
-
-// Personalized Hook
-
-export function useTaskContext() {
-  return useContext(TaskContext);
-}
