+++ src/contexts/TaskContext/useTaskContext.ts
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import { TaskContext } from '.';
+
+// Personalized Hook
+
+export function useTaskContext() {
+  return useContext(TaskContext);
+}
